+++ FamilyAccounting.DAL/Repositories/TransactionRepository.cs
@@ -25,7 +25,7 @@ namespace FamilyAccounting.DAL.Repositories
                 };
                 command.Parameters.AddWithValue("@_id_wallet", transaction.SourceWalletId);
                 command.Parameters.AddWithValue("@_amount", transaction.Amount);
-                //command.Parameters.AddWithValue("@_id_category", transaction.Category.Id);
+                command.Parameters.AddWithValue("@_id_category", transaction.CategoryId);
                 command.Parameters.AddWithValue("@_description", transaction.Description);
                 SqlParameter output = new SqlParameter
                 {
@@ -44,8 +44,7 @@ namespace FamilyAccounting.DAL.Repositories
         {
             using (SqlConnection con = new SqlConnection(connectionString))
             {
-                //string sqlExpression = $"EXEC PR_Actions_Update {id}, '{transaction.Category.Id}', '{transaction.Description}'";
-                string sqlExpression = $"EXEC";
+                string sqlExpression = $"EXEC PR_Actions_Update {id}, '{transaction.CategoryId}', '{transaction.Description}'";
                 SqlConnection sql = new SqlConnection(connectionString);
                 SqlCommand command = new SqlCommand(sqlExpression, sql);
                 if (con.State != ConnectionState.Open)
@@ -78,7 +77,7 @@ namespace FamilyAccounting.DAL.Repositories
                     {
                         var t = new Transaction();
                         t.Id = dr.GetInt32("id");
-                        //t.Category.Id = dr.GetInt32("id_category");
+                        t.CategoryId = dr.GetInt32("id_category");
                         t.Description = dr.GetString("description");
                         transaction = t;
                     }
