+++ FamilyAccounting.Web/Views/Transaction/MakeExpense.cshtml
@@ -7,14 +7,14 @@
 <h1>Make expense</h1>
 
 <div>
-    @*<form method="post" asp-controller="Transaction" asp-action="MakeExpense">
-        <input placeholder="Enter id of a source wallet" style="display:none" asp-for="SourceWallet.Id" />
+    <form method="post" asp-controller="Transaction" asp-action="MakeExpense">
+        <input placeholder="Enter id of a source wallet" style="display:none" asp-for="SourceWalletId" />
         <input placeholder="Describe the expense" asp-for="Description" />
         <input placeholder="Enter amount" asp-for="Amount" />
-        <input placeholder="Enter the id of category" asp-for="Category.Id" />
+        <input placeholder="Enter the id of category" asp-for="CategoryId" />
         <button type="submit" asp-controller="Transaction" asp-action="MakeExpense">
             Make expense
         </button>
-        <button asp-controller="Wallet" asp-action="Details" asp-route-id="@Model.SourceWallet.Id">Back to wallet</button>
-    </form>*@
+        <button asp-controller="Wallet" asp-action="Details" asp-route-id="@Model.SourceWalletId">Back to wallet</button>
+    </form>
 </div>
