+++ FamilyAccounting.Web/Views/Transaction/Update.cshtml
@@ -17,11 +17,11 @@
                 <input asp-for="Description" class="form-control" />
                 <span asp-validation-for="Description" class="text-danger"></span>
             </div>
-            @*<div class="form-group">
-                <label asp-for="Category.Id" class="control-label"></label>
-                <input asp-for="Category.Id" class="form-control" />
+            <div class="form-group">
+                <label asp-for="CategoryId" class="control-label"></label>
+                <input asp-for="CategoryId" class="form-control" />
                 <span asp-validation-for="Id" class="text-danger"></span>
-            </div>*@
+            </div>
             <div class="form-group">
                 <input type="submit" value="Save" class="btn btn-primary" />
             </div>
