+++ FamilyAccounting.Web/Views/Wallet/Create.cshtml
@@ -12,8 +12,8 @@
         <form asp-action="Create">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <div class="form-group" style="display:none">
-                @*<label asp-for="Person.Id" class="control-label"></label>
-                <input asp-for="Person.Id" class="form-control" />*@
+                <label asp-for="PersonId" class="control-label"></label>
+                <input asp-for="PersonId" class="form-control" />
                 <span asp-validation-for="Id" class="text-danger"></span>
             </div>
             <div class="form-group">
